Factorize some methods in DownloadWorkerUtils
